Fix file uploads to google cloud storage
Fix uploads of file to google cloud storage by replacing
storage helper class to point to Google Cloud instead of the user disk.

Add a middleware to handle the upload to google bucket. This makes the
last handler to handler only creating of the todo item and returning a
response to the user.

[Ticket: X]